+++ app/src/main/AndroidManifest.xml
@@ -22,7 +22,7 @@
         tools:targetApi="31">
 
         <activity
-            android:name=".RegisterCareActivity"
+            android:name=".ui.mypage.detail.MyPageDetailActivity"
             android:exported="false" />
         <activity
             android:name=".ui.main.MainActivity"
@@ -30,8 +30,8 @@
             android:windowSoftInputMode="adjustPan" />
         <activity
             android:name="org.retriever.dailypet.ui.login.LoginActivity"
-            android:windowSoftInputMode="adjustPan"
-            android:exported="true">
+            android:exported="true"
+            android:windowSoftInputMode="adjustPan">
             <intent-filter android:autoVerify="true">
                 <action android:name="android.intent.action.VIEW" />
 
@@ -49,9 +49,9 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".ui.mypage.WebViewActivity"
+            android:name=".ui.mypage.webview.WebViewActivity"
             android:exported="false" />
-              <activity
+        <activity
             android:name=".ui.splash.SplashActivity"
             android:exported="true"
             android:theme="@style/Splash">
@@ -76,6 +76,7 @@
                     android:scheme="${KAKAO_NATIVE_APP_KEY}" />
             </intent-filter>
         </activity>
+
         <service
             android:name=".interfaces.MyFirebaseMessagingService"
             android:exported="true">
