+++ app/src/main/java/org/retriever/dailypet/ui/login/LoginActivity.kt
@@ -9,10 +9,5 @@ import org.retriever.dailypet.ui.base.BaseActivity
 
 @AndroidEntryPoint
 class LoginActivity : BaseActivity<ActivityLoginBinding>({ ActivityLoginBinding.inflate(it) }) {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-    }
-
 }
