+++ app/src/main/java/org/retriever/dailypet/ui/mypage/MyPageMainFragment.kt
@@ -18,7 +18,6 @@ import org.retriever.dailypet.R
 import org.retriever.dailypet.databinding.FragmentMyPageMainBinding
 import org.retriever.dailypet.model.Resource
 import org.retriever.dailypet.ui.base.BaseFragment
-import org.retriever.dailypet.ui.bottomsheet.GroupBottomSheet
 import org.retriever.dailypet.ui.login.LoginActivity
 import org.retriever.dailypet.util.hideProgressCircular
 import org.retriever.dailypet.util.showProgressCircular
@@ -48,12 +47,8 @@ class MyPageMainFragment : BaseFragment<FragmentMyPageMainBinding>() {
     }
 
     private fun buttonClick() = with(binding) {
-        managePetText.setOnClickListener {
-            root.findNavController().navigate(R.id.action_myPageMainFragment_to_myPagePetFragment)
-        }
-
         manageGroupText.setOnClickListener {
-            showGroupBottomSheet()
+            root.findNavController().navigate(R.id.action_myPageMainFragment_to_myPageDetailActivity)
         }
 
         groupInviteText.setOnClickListener {
@@ -101,11 +96,6 @@ class MyPageMainFragment : BaseFragment<FragmentMyPageMainBinding>() {
             val intent = Intent(requireContext(), LoginActivity::class.java)
             startActivity(intent)
         }
-    }
-
-    private fun showGroupBottomSheet() {
-        val groupBottomSheet = GroupBottomSheet()
-        groupBottomSheet.show(childFragmentManager, groupBottomSheet.tag)
     }
 
     private fun logout() {
