+++ app/src/main/java/org/retriever/dailypet/ui/mypage/MyPageViewModel.kt
@@ -20,32 +20,10 @@ class MyPageViewModel @Inject constructor(
     private val _withdrawalResponse = MutableLiveData<Resource<ResponseBody>>()
     val withdrawalResponse: LiveData<Resource<ResponseBody>> = _withdrawalResponse
 
-    private val _petDetailResponse = MutableLiveData<Resource<PetDetailResponse>>()
-    val petDetailResponse: LiveData<Resource<PetDetailResponse>> = _petDetailResponse
-
-    private val _deletePetResponse = MutableLiveData<Resource<ResponseBody>>()
-    val deletePetResponse: LiveData<Resource<ResponseBody>> = _deletePetResponse
-
     fun deleteMemberWithdrawal(jwt: String) = viewModelScope.launch {
         _withdrawalResponse.postValue(Resource.Loading())
 
         _withdrawalResponse.postValue(myPageRepository.deleteMemberWithdrawal(jwt))
-    }
-
-    fun getPetList(familyId: Int, jwt: String) {
-        viewModelScope.launch {
-            _petDetailResponse.postValue(Resource.Loading())
-
-            _petDetailResponse.postValue(myPageRepository.getPetList(familyId, jwt))
-        }
-    }
-
-    fun deletePet(jwt: String, familyId: Int, petId: Int) {
-        viewModelScope.launch {
-            _deletePetResponse.postValue(Resource.Loading())
-
-            _deletePetResponse.postValue(myPageRepository.deletePet(jwt, familyId, petId))
-        }
     }
 
 }
