+++ app/src/main/java/org/retriever/dailypet/ui/mypage/adapter/GroupAdapter.kt
@@ -5,19 +5,20 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
-import org.retriever.dailypet.databinding.ItemFamilyBinding
+import org.retriever.dailypet.databinding.ItemMyPageDetailBinding
+import org.retriever.dailypet.databinding.ItemProfileBinding
 import org.retriever.dailypet.model.mypage.GroupDetailItem
 
 class GroupAdapter : ListAdapter<GroupDetailItem, GroupAdapter.ViewHolder>(diffUtil) {
 
-    class ViewHolder(val binding: ItemFamilyBinding) : RecyclerView.ViewHolder(binding.root) {
+    class ViewHolder(val binding: ItemMyPageDetailBinding) : RecyclerView.ViewHolder(binding.root) {
         fun bind(item: GroupDetailItem) {
             binding.profileNickName.text = item.groupRoleName
         }
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-        return ViewHolder(ItemFamilyBinding.inflate(LayoutInflater.from(parent.context), parent, false))
+        return ViewHolder(ItemMyPageDetailBinding.inflate(LayoutInflater.from(parent.context), parent, false))
     }
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
