+++ app/src/main/java/org/retriever/dailypet/ui/mypage/detail/MyPageDetailActivity.kt
@@ -1,0 +1,9 @@
+package org.retriever.dailypet.ui.mypage.detail
+
+import dagger.hilt.android.AndroidEntryPoint
+import org.retriever.dailypet.databinding.ActivityMyPageDetailBinding
+import org.retriever.dailypet.ui.base.BaseActivity
+
+@AndroidEntryPoint
+class MyPageDetailActivity : BaseActivity<ActivityMyPageDetailBinding>({ ActivityMyPageDetailBinding.inflate(it) }) {
+}
