+++ app/src/main/java/org/retriever/dailypet/ui/signup/adapter/FindGroupAdapter.kt
@@ -4,20 +4,20 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import org.retriever.dailypet.R
-import org.retriever.dailypet.databinding.ItemFamilyBinding
+import org.retriever.dailypet.databinding.ItemProfileBinding
 
 class FindGroupAdapter(
     private val familyList: List<String>
 ) : RecyclerView.Adapter<FindGroupAdapter.ViewHolder>() {
 
-    inner class ViewHolder(val binding: ItemFamilyBinding) : RecyclerView.ViewHolder(binding.root) {
+    inner class ViewHolder(val binding: ItemProfileBinding) : RecyclerView.ViewHolder(binding.root) {
         val nickName = binding.profileNickName
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         val inflater = LayoutInflater.from(parent.context)
-        val view = inflater.inflate(R.layout.item_family, parent, false)
-        return ViewHolder(ItemFamilyBinding.bind(view))
+        val view = inflater.inflate(R.layout.item_profile, parent, false)
+        return ViewHolder(ItemProfileBinding.bind(view))
     }
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
