+++ app/src/main/res/layout/fragment_create_family.xml
@@ -36,7 +36,7 @@
         android:layout_height="wrap_content"
         android:layout_marginTop="24dp"
         android:background="@android:color/transparent"
-        android:src="@drawable/icon_back"
+        app:srcCompat="@drawable/arrow_left"
         app:layout_constraintStart_toStartOf="@id/left_guideline"
         app:layout_constraintTop_toTopOf="parent" />
