+++ app/src/main/res/navigation/my_page_detail_navigation_graph.xml
@@ -1,0 +1,43 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/my_page_detail_navigation_graph"
+    app:startDestination="@id/myPageDetailFragment">
+
+    <fragment
+        android:id="@+id/myPageDetailFragment"
+        android:name="org.retriever.dailypet.ui.mypage.detail.MyPageDetailFragment"
+        android:label="MyPageDetailFragment" >
+        <action
+            android:id="@+id/action_myPageDetailFragment_to_myPagePetDetailFragment"
+            app:destination="@id/myPagePetDetailFragment" />
+        <action
+            android:id="@+id/action_myPageDetailFragment_to_createPetFragment2"
+            app:destination="@id/createPetFragment2" />
+    </fragment>
+    <fragment
+        android:id="@+id/myPagePetDetailFragment"
+        android:name="org.retriever.dailypet.ui.mypage.detail.MyPagePetDetailFragment"
+        android:label="MyPagePetDetailFragment" >
+        <argument
+            android:name="petDetailItem"
+            app:argType="org.retriever.dailypet.model.mypage.PetDetailItem" />
+        <action
+            android:id="@+id/action_myPagePetDetailFragment_to_createPetFragment2"
+            app:destination="@id/createPetFragment2" />
+    </fragment>
+    <fragment
+        android:id="@+id/createPetFragment2"
+        android:name="org.retriever.dailypet.ui.signup.CreatePetFragment"
+        android:label="CreatePetFragment" >
+        <argument
+            android:name="isAdd"
+            app:argType="boolean"
+            android:defaultValue="false" />
+        <argument
+            android:name="petDetailItem"
+            app:argType="org.retriever.dailypet.model.mypage.PetDetailItem"
+            app:nullable="true"
+            android:defaultValue="@null" />
+    </fragment>
+</navigation>
